+++ changes/24771-mdm-deadlock-fixes
@@ -1,0 +1,1 @@
+Fixed potential deadlocks when deploying Apple configuration profiles.
